PB-114: Moved utils function into utils
