+++ src/modules/drawing/lib/drawingUtils.js
@@ -1,18 +1,6 @@
 import { wrapX } from 'ol/coordinate'
-import { LineString, Point, Polygon } from 'ol/geom'
+import { LineString, Polygon } from 'ol/geom'
 import { get as getProjection } from 'ol/proj'
-import proj4 from 'proj4'
-
-import { LV95 } from '@/utils/coordinates/coordinateSystems'
-import { format } from '@/utils/numberUtils'
-
-export function toLv95(input, epsg) {
-    if (Array.isArray(input[0])) {
-        return input.map((si) => toLv95(si, epsg))
-    } else {
-        return proj4(epsg, LV95.epsg, [input[0], input[1]])
-    }
-}
 
 /**
  * Wraps the provided coordinates in the world extents (i.e. the coordinate range that if equivalent
@@ -33,72 +21,6 @@ export function wrapXCoordinates(coordinates, projection, inPlace = false) {
         return wrappedCoords.map((c) => wrapXCoordinates(c, projection, inPlace))
     }
     return wrapX(wrappedCoords, getProjection(projection.epsg))
-}
-
-/** @param {[number, number]} coo Coordinates */
-export function formatPointCoordinates(coo) {
-    return `${format(coo[0], 0)}, ${format(coo[1], 0)}`
-}
-
-export function formatMeters(value, { dim = 1, digits = 2, applyFormat = true } = {}) {
-    const factor = Math.pow(1000, dim)
-    let unit = dim === 1 ? 'm' : 'm²'
-    if (value >= factor) {
-        unit = dim === 1 ? 'km' : 'km²'
-        value /= factor
-    }
-    value = applyFormat ? format(value, digits) : value.toFixed(digits)
-    return `${value} ${unit}`
-}
-
-export function geometryInfo(type, coordinates, epsg) {
-    const coos95 = toLv95(coordinates, epsg)
-    const output = {
-        type,
-    }
-    if (type === Point) {
-        output.location = formatPointCoordinates(coos95)
-    } else {
-        if (type === Polygon) {
-            const poly = new Polygon(coos95)
-            output.area = formatMeters(poly.getArea(), { dim: 2 })
-            const line = new LineString(coos95[0])
-            output.perimeter = formatMeters(line.getLength())
-        } else {
-            const line = new LineString(coos95)
-            output.length = formatMeters(line.getLength())
-        }
-    }
-    return output
-}
-
-/**
- * Formats minutes to hours and minutes (if more than one hour) e.g. 1230 -> '20h 30min', 55 ->
- * '55min'
- *
- * @param {Number} minutes
- * @returns {string}
- */
-export function formatTime(minutes) {
-    if (isNaN(minutes) || minutes === null) {
-        return '-'
-    }
-    let result = ''
-    if (minutes >= 60) {
-        let hours = Math.floor(minutes / 60)
-        minutes = minutes - hours * 60
-        result += `${hours}h`
-        if (minutes > 0) {
-            result += ` ${minutes}min`
-        }
-    } else {
-        result += `${minutes}min`
-    }
-    return result
-}
-
-export function formatAngle(value, digits = 2) {
-    return `${value.toFixed(digits)}°`
 }
 
 /**
