+++ src/utils/geodesicManager.js
@@ -10,9 +10,9 @@ import RBush from 'ol/structs/RBush' /* Warning: private class of openlayers */
 import { Circle, Fill, RegularShape, Stroke, Style, Text } from 'ol/style'
 import proj4 from 'proj4'
 
-import { formatAngle, formatMeters } from '@/modules/drawing/lib/drawingUtils'
 import { WGS84 } from '@/utils/coordinates/coordinateSystems'
 import log from '@/utils/logging'
+import { formatAngle, formatMeters } from '@/utils/utils'
 
 const geod = Geodesic.WGS84
