+++ src/utils/utils.js
@@ -1,4 +1,8 @@
+import { LineString, Point, Polygon } from 'ol/geom'
+
+import { toLv95 } from '@/utils/coordinates/coordinateUtils'
 import log from '@/utils/logging'
+import { format } from '@/utils/numberUtils'
 
 /**
  * Check if the provided string is a valid URL
@@ -81,4 +85,92 @@ export function generateFilename(fileExtension) {
         .replaceAll(':', '')
         .replace('T', '')
     return `map.geo.admin.ch_${fileExtension.toUpperCase()}_${date}.${fileExtension.toLowerCase()}`
+}
+
+/**
+ * Formats minutes to hours and minutes (if more than one hour) e.g. 1230 -> '20h 30min', 55 ->
+ * '55min'
+ *
+ * @param {Number} minutes
+ * @returns {string} Time in 'Hh Mmin'
+ */
+export function formatMinutesTime(minutes) {
+    if (isNaN(minutes) || minutes === null) {
+        return '-'
+    }
+    let result = ''
+    if (minutes >= 60) {
+        let hours = Math.floor(minutes / 60)
+        minutes = minutes - hours * 60
+        result += `${hours}h`
+        if (minutes > 0) {
+            result += ` ${minutes}min`
+        }
+    } else {
+        result += `${minutes}min`
+    }
+    return result
+}
+
+/**
+ * Format Coordinates
+ *
+ * @param {[number, number]} coo Coordinates
+ * @returns {String} Returns coordinate in a readable format without decimals
+ */
+export function formatPointCoordinates(coo) {
+    return `${format(coo[0], 0)}, ${format(coo[1], 0)}`
+}
+
+/**
+ * Format distance or ara in a readable format
+ *
+ * @param {Number} value
+ * @param {{ dim: Number; digits: Number; applyFormat: Boolean }} options
+ * @returns {String} Distance/area formatted (e.g. 1000 => '1 km')
+ */
+export function formatMeters(value, { dim = 1, digits = 2, applyFormat = true } = {}) {
+    const factor = Math.pow(1000, dim)
+    let unit = dim === 1 ? 'm' : 'm²'
+    if (value >= factor) {
+        unit = dim === 1 ? 'km' : 'km²'
+        value /= factor
+    }
+    value = applyFormat ? format(value, digits) : value.toFixed(digits)
+    return `${value} ${unit}`
+}
+
+/**
+ * Format a geometry info in human readable format
+ *
+ * WARNING the Coordinate is reprojected to LV95
+ *
+ * @param {any} type
+ * @param {any} coordinates
+ * @param {any} epsg
+ * @returns {String}
+ */
+export function geometryInfo(type, coordinates, epsg) {
+    const coos95 = toLv95(coordinates, epsg)
+    const output = {
+        type,
+    }
+    if (type === Point) {
+        output.location = formatPointCoordinates(coos95)
+    } else {
+        if (type === Polygon) {
+            const poly = new Polygon(coos95)
+            output.area = formatMeters(poly.getArea(), { dim: 2 })
+            const line = new LineString(coos95[0])
+            output.perimeter = formatMeters(line.getLength())
+        } else {
+            const line = new LineString(coos95)
+            output.length = formatMeters(line.getLength())
+        }
+    }
+    return output
+}
+
+export function formatAngle(value, digits = 2) {
+    return `${value.toFixed(digits)}°`
 }
